requirements.txt
# Core dependencies
elasticsearch>=8.0.0
requests>=2.25.0

# Optional: Load configuration from .env file
python-dotenv>=0.19.0
